+++ pages/_app.tsx
@@ -2,7 +2,6 @@
 import type { AppProps } from "next/app";
 import "../styles/navbar.css";
 import "../styles/auth.css";
-import "../styles/activity.css";
 import "../styles/view.css";
 import "../styles/edit.css";
 import "../styles/sidebar.css";
