+++ pages/index.tsx
@@ -1,12 +1,12 @@
 import { Inter } from "@next/font/google";
-import Navbar from "./navbar";
+import LoginPage from "./loginPage";
 
 const inter = Inter({ subsets: ["latin"] });
 
 export default function Home() {
   return (
     <>
-      <Navbar />
+      <LoginPage />
     </>
   );
 }
