--- styles/activity.css
@@ -1,3 +1,0 @@
-.err_str {
-  color: red;
-}
