--- styles/globals.css
@@ -1,86 +1,0 @@
-header {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 30px;
-}
-.nav-bar button {
-  background-color: "#26ceb3";
-  color: white;
-  font-size: 17px;
-  font-weight: 900;
-  padding-left: 25px;
-  padding-right: 25px;
-}
-
-.intro img,
-.productive img {
-  width: 100%;
-  max-width: 600px;
-  height: auto;
-}
-
-button {
-  border-radius: 25px;
-  border: none;
-  height: 3rem;
-  font-weight: 700;
-  padding: 0 12px;
-}
-
-.btn {
-  width: 15rem;
-  max-width: 80%;
-}
-
-.productive {
-  display: flex;
-  align-items: center;
-}
-
-.productive-content {
-  padding-left: 7rem;
-}
-
-@media (max-width: 1000px) {
-  header img,
-  footer img {
-    width: 117px;
-    height: auto;
-  }
-  .intro > div,
-  .features-item {
-    width: 100%;
-  }
-  .productive {
-    flex-direction: column;
-  }
-}
-
-@media (max-width: 375px) {
-  .nav-bar button {
-    background-color: "#26ceb3";
-    color: white;
-    font-size: 8px;
-    font-weight: 700;
-    padding-left: 20px;
-    padding-right: 20px;
-  }
-}
-
-footer {
-  text-align: left;
-  margin-top: 70px;
-}
-
-footer p {
-  color: rgba(12, 20, 15, 1);
-  font-size: 16px;
-  letter-spacing: 0.4px;
-}
-
-footer a {
-  color: #4b4d50;
-  text-decoration: none;
-  transition: all 0.2s ease;
-}
-
